get_spotify_audio_features_data_to_bigquery/requirements.txt
spotipy
google-cloud-bigquery